requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
streamlit==1.32.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.20.0
transformers==4.40.0
torch==2.3.0
db-dtypes==1.2.0